separando endereço de usuario
