+++ src/main/java/model/vo/Endereco.java
@@ -1,0 +1,93 @@
+package model.vo;
+
+public class Endereco {
+	private String rua;
+	private int numeroRua;
+	private String bairro;
+	private String cidade;
+	private String uf;
+	private String cep;
+	
+	
+	public Endereco(String rua, int numeroRua, String bairro, String cidade, String uf, String cep) {
+		super();
+		this.rua = rua;
+		this.numeroRua = numeroRua;
+		this.bairro = bairro;
+		this.cidade = cidade;
+		this.uf = uf;
+		this.cep = cep;
+	}
+	
+	
+
+
+	public Endereco() {
+		super();
+	}
+
+
+
+
+	public String getRua() {
+		return rua;
+	}
+
+
+	public void setRua(String rua) {
+		this.rua = rua;
+	}
+
+
+	public int getNumeroRua() {
+		return numeroRua;
+	}
+
+
+	public void setNumeroRua(int numeroRua) {
+		this.numeroRua = numeroRua;
+	}
+
+
+	public String getBairro() {
+		return bairro;
+	}
+
+
+	public void setBairro(String bairro) {
+		this.bairro = bairro;
+	}
+
+
+	public String getCidade() {
+		return cidade;
+	}
+
+
+	public void setCidade(String cidade) {
+		this.cidade = cidade;
+	}
+
+
+	public String getUf() {
+		return uf;
+	}
+
+
+	public void setUf(String uf) {
+		this.uf = uf;
+	}
+
+
+	public String getCep() {
+		return cep;
+	}
+
+
+	public void setCep(String cep) {
+		this.cep = cep;
+	}
+	
+	
+	
+}
